Add files via upload
This program is to map the single-phase homogeneity range of an extended phase in a quaternary system and to make a 3D diagram. 

To run this program, one needs
(1) Thermo-Calc Software package, with TC-Python API
(2) A thermodynamic database containing more than 4 components

This program is to map the single-phase homogeneity range of an extended phase in a quaternary
system.